Release 0.0.5

+++ install.sh
@@ -7,8 +7,12 @@ bin="/usr/local/bin"
 # Copy the path
 sudo cp -R $dirname "$lib/"
 
+# remove existing bin if it exists
+if [ -e "$bin/dots" ]; then
+  rm "$bin/dots"
+fi
+
 # symlink dots
-rm "$bin/dots"
 ln -s "$lib/dots/dots.sh" "$bin/dots"
 
 # Ubuntu-only: Change from dash to bash
